chore: move conversation helper function to ModelHelper
